Frederic Bouvier:
Well, I found a bug in VpfTile that prevents the function getTileSubdir
to return the correct value.

In that case getTileSubdir was returning the content of a local
variable (subdir) that is destroyed before used outside. Usually,
it generates a segfault with MSVC but here, the returned string
was "". The patch returns the value of the string and tgpvf is
going well so far.