Fred Bouvier:
I replaced the test of the presence of PUI by a test of the presence of SG, and
I moved the C++ instruction to include PLIB lib check ( on UL ). PUI is not
used by TerraGear and it uses OpenGL that I don't want under cygwin because I
would have to install X11 and it is a mess, just for text tools.