Added new texturing method. Added sanity check for polygon file reading.
New texturing method TG_TEX_HORIZ_REF requires that each polygon
has a reference texture coordinate specified for a specific point on the
polygon. The coordinate changes parallel to a direction (also
stored) and the other texture coordinate changes perpendicular
to this direction. The reference texture coordinates in polygons
of the same type can be chosen to match that of their neighbours
to produce a continous map onto a texture. This technique can
be used to texture rivers or cliffs as their direction
changes.

The sanity checks simply make sure that the number of contours
or polygons reported in a file are not too large. This could
happen if a polygon file was generated by a previous version
of terragear tools with an incompatible layout of polygon
information, or a non-polygon file is present in the input
directory and is opened as a polygon file.